services/semantic-search/requirements.txt
chromadb
google-cloud-pubsub
